test: Assembler is no longer dependent on weaver
