+++ Assets/Tests/Weaver/CompiledAssembly.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 37e2f3c74223040efa6e0c5b26f1d833
